Making images fit on mobile
I'm only doing this for 1st and 2nd page for now, as older
posts were written in Blogger and require a bit more work
and they most likely won't be read in the future, even less on
mobile. The very important thing to guarantee now is that the front
page, in the present and the future, will also correctly fit on
mobile.