Simply title of last post
I was going to just change 1 word (at the end of the title) between
HomeDC 02 and HomeDC 03 but I wasn't happy with it after writing HomeDC 03.

Unfortunately Google has already indexed the old title/URL but hopefully
it will soon be removed and replaced by the new one this commit introduces.